Initial commit of RGW automation
- Rename mcg_bucket.py to objectbucket.py in order to be more general
- Move the OBC class from bucket_policy.py to objectbucket.py
- Move verify_s3_object_integrity from mcg to helpers
- Adjust the OBC class init logic to fit both RGW and MCG OBCs
- Add a new RGW class to represent RGW in the cluster
- Add independent mode RGW SC name to constants
- Edit the OBC YAML to more easily fit into RGW OC flow
- Refactor sync_object_directory to not rely on mcg_obj
- Rename classes in objectbucket.py to be more representative
- Refactor bucket_factory to work with both RGW and MCG
- Refactor awscli_pod to not rely on mcg_obj
- Add RGW bucket creation test
- Add RGW object integrity test
