+++ ocs_ci/ocs/constants.py
@@ -172,6 +172,10 @@ LOCAL_STORAGE_NAMESPACE = 'local-storage'
 DEFAULT_STORAGECLASS_CEPHFS = f'{DEFAULT_CLUSTERNAME}-cephfs'
 DEFAULT_STORAGECLASS_RBD = f'{DEFAULT_CLUSTERNAME}-ceph-rbd'
 
+# Independent mode default StorageClasses
+INDEPENDENT_DEFAULT_CLUSTER_NAME = 'ocs-independent-storagecluster'
+INDEPENDENT_DEFAULT_STORAGECLASS_RGW = f'{INDEPENDENT_DEFAULT_CLUSTER_NAME}-ceph-rgw'
+
 # encoded value of 'admin'
 ADMIN_USER = 'admin'
 GB = 1024 ** 3
