+++ tests/helpers.py
@@ -8,6 +8,7 @@ import json
 import logging
 import os
 import re
+import shlex
 import statistics
 import tempfile
 import threading
@@ -2330,3 +2331,28 @@ def retrieve_default_ingress_crt():
 
     with open(constants.DEFAULT_INGRESS_CRT_LOCAL_PATH, 'w') as crtfile:
         crtfile.write(decoded_crt)
+
+
+def verify_s3_object_integrity(original_object_path, result_object_path, awscli_pod):
+    """
+    Verifies checksum between orignial object and result object on an awscli pod
+
+    Args:
+        original_object_path (str): The Object that is uploaded to the s3 bucket
+        result_object_path (str):  The Object that is downloaded from the s3 bucket
+        awscli_pod (pod): A pod running the AWSCLI tools
+
+    Returns:
+            bool: True if checksum matches, False otherwise
+
+    """
+    md5sum = shlex.split(awscli_pod.exec_cmd_on_pod(command=f'md5sum {original_object_path} {result_object_path}'))
+    if md5sum[0] == md5sum[2]:
+        logger.info(f'Passed: MD5 comparison for {original_object_path} and {result_object_path}')
+        return True
+    else:
+        logger.error(
+            f'Failed: MD5 comparison of {original_object_path} and {result_object_path} - '
+            f'{md5sum[0]} ≠ {md5sum[2]}'
+        )
+        return False
