+++ tests/manage/mcg/test_multi_region.py
@@ -11,6 +11,7 @@ from ocs_ci.ocs.constants import BS_AUTH_FAILED, BS_OPTIMAL
 from ocs_ci.ocs.exceptions import TimeoutExpiredError
 from ocs_ci.utility.utils import TimeoutSampler
 from tests.manage.mcg.helpers import retrieve_test_objects_to_pod, sync_object_directory
+from tests.helpers import verify_s3_object_integrity
 
 logger = logging.getLogger(__name__)
 
@@ -99,7 +100,7 @@ class TestMultiRegion:
 
         # Checksum is compared between original and result object
         for obj in downloaded_objs:
-            assert mcg_obj.verify_s3_object_integrity(
+            assert verify_s3_object_integrity(
                 original_object_path=f'{local_testobjs_dir_path}/{obj}',
                 result_object_path=f'{local_temp_path}/{obj}', awscli_pod=awscli_pod
             ), 'Checksum comparision between original and result object failed'
@@ -121,7 +122,7 @@ class TestMultiRegion:
 
         # Checksum is compared between original and result object
         for obj in downloaded_objs:
-            assert mcg_obj.verify_s3_object_integrity(
+            assert verify_s3_object_integrity(
                 original_object_path=f'{local_testobjs_dir_path}/{obj}',
                 result_object_path=f'{local_temp_path}/{obj}', awscli_pod=awscli_pod
             ), 'Checksum comparision between original and result object failed'
