+++ tests/manage/mcg/test_object_integrity.py
@@ -6,6 +6,7 @@ from ocs_ci.framework.testlib import ManageTest, tier1, tier2, tier3
 from ocs_ci.ocs import constants
 from tests.manage.mcg import helpers
 from tests.manage.mcg.helpers import retrieve_anon_s3_resource
+from tests.helpers import verify_s3_object_integrity
 
 logger = logging.getLogger(__name__)
 
@@ -48,7 +49,7 @@ class TestObjectIntegrity(ManageTest):
 
         # Checksum is compared between original and result object
         for obj in downloaded_files:
-            assert mcg_obj.verify_s3_object_integrity(
+            assert verify_s3_object_integrity(
                 original_object_path=f'{original_dir}/{obj}',
                 result_object_path=f'{result_dir}/{obj}', awscli_pod=awscli_pod
             ), 'Checksum comparision between original and result object failed'
@@ -136,7 +137,7 @@ class TestObjectIntegrity(ManageTest):
 
             # Checksum is compared between original and result object
             for obj in download_files:
-                assert mcg_obj.verify_s3_object_integrity(
+                assert verify_s3_object_integrity(
                     original_object_path=f'{original_dir}/{obj}',
                     result_object_path=f'{result_dir}/{obj}',
                     awscli_pod=awscli_pod
@@ -178,7 +179,7 @@ class TestObjectIntegrity(ManageTest):
 
         # Checksum is compared between original and result object
         for obj in ('test' + str(i + 1) for i in range(100)):
-            assert mcg_obj.verify_s3_object_integrity(
+            assert verify_s3_object_integrity(
                 original_object_path=f'{original_dir}/{obj}',
                 result_object_path=f'{result_dir}/{obj}', awscli_pod=awscli_pod
             ), 'Checksum comparision between original and result object failed'
